Switched from `ifconfig | grep` to system methods (#216)
for retrieving interfaces MAC addresses to reduce conflicts between versions of ifconfig.
